pyproject.toml
[project]
name = "learnify"
version = "0.1.0"
description = "Learnify is a smart, Gemini-powered multi-agent assistant designed to help students."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
